Reorganize the project
 - separate file for builtin function
 - updated README.md
