+++ src/builtin.h
@@ -1,0 +1,10 @@
+//
+// Created by duskybomb on 11/11/18.
+//
+
+#ifndef SHELL_BUILTIN_H
+#define SHELL_BUILTIN_H
+
+extern char *builtin_str[];
+
+#endif //SHELL_BUILTIN_H
